Modificar o app.js para reconhecer a rota telefones; pequenas modificações para organizar os outros arquivos

+++ back_end/src/routes/cidade.js
@@ -31,7 +31,7 @@ router.post('/especifica',usuario,(req,res,next)=>{
     pool.getConnection((error,conn)=>{
 
         if(!req.body.id_estado){res.status(406).send({mensagem:"Estado necessário"})}
-        
+
         if(error){return res.status(500).send({error:error})}
         conn.query('SELECT * FROM cidade WHERE id_estado=? ',[req.body.id_estado],(error,results,field)=>{
             if(error){return res.status(500).send({error:error})}
@@ -154,6 +154,7 @@ router.patch('/',mestre,(req,res,next)=>{
  */
 router.delete('/',gerente,(req,res,next)=>{
     pool.getConnection((error,conn)=>{
+        if(!req.body.id_cidade){return res.status(406).send("Insira o número identificador da cidade")}
         if(error){return res.status(500).send({error:error})}
         conn.query(
             'DELETE FROM cidade WHERE id_cidade = ?',
